+++ examples/encode_decode_simple/encode_decode_simple.c
@@ -35,11 +35,11 @@ int main()
     uint32_t max_symbol_size = 160;
 
     // Here we select the coding algorithm we wish to use
-    size_t algorithm = kodo_full_rlnc;
+    int32_t algorithm = kodo_full_rlnc;
 
     // Here we select the finite field to use common choices are
     // kodo_binary, kodo_binary8, kodo_binary16
-    size_t finite_field = kodo_binary;
+    int32_t finite_field = kodo_binary;
 
     kodo_factory_t* encoder_factory =
         kodo_new_encoder_factory(algorithm, finite_field,
